Warning bugs and variable name changes
[changes]
+Change abstract variable name 'n' for 'some_node'.
-Comment all bug functions and delete useless things.

[need]
+Need rebuild: 'erase' and all 'node' functions.

[comments]
I discovered the reason with the remove_node problem:
The list was compressed because of the bad pointer manipulation.
An example of this error is 'erase' function definition.

[destroyed-warning]
DON'T DO THIS FOR WALK:
*some_node = (*some_node)->next;

In this moment, I get each pointer of the links and swap for next.
The final result it's:

"the first node equal to the final node that the execution arrived."

In other words: I LOSE [0, n] ELEMENTS!