Final implementations and fixes! Finally!!! : D
[change]
+erase =: rebuild
+erase_search := create
+tail_pointer := create
+insert_node := create
+remove_node := fixes
+insert (queue, stack, deque) := recursive algorithm
+pop (queue, stack, deque) := delicious method of remove_node
-anti-delicious-features := thrown into some black hole

[comments]
I finally finished this work.
This final implementation was very bloody.
The way to walk on linked lists getting the
correct references (pointer-to-pointer) is not easy.
But now I'm deliciously happy.
Final result are very good.
And after all the destruction on compiler and debugger:
appears that It was worth all the effort.

[messages-for-future-generation]
Caution! Very caution with manipulate pointers!
Use and abuse of debugger: its your friend!
And now it's my also.
We made a beautiful friendship in this repository. <3

[recommendations-and-tips]
Listen Muse!
They'll inspire you (maybe even give solutions), we do well
and still give strength to continue!

[my-insight-font]
Very good exercises:
https://www.hackerrank.com/domains/data-structures/linked-lists
I did some exercises of this set and helped me so much.
Try do these exercises! Worth it.